+++ src/com/company/Shopping/Silpo.java
@@ -12,7 +12,7 @@ public class Silpo implements Megamoll {
     }
 
     @Override
-    public void addInListPurchasingCharacteristics(String purchaseName, double purchasePrise) {
+    public void addInListPurchasingCharacteristics(String purchaseName, int purchasePrise) {
         productsSilpo.add(new PurchasingCharacteristics(purchaseName, purchasePrise));
     }
